+++ spec/table_structure/schema/definition/compiler_spec.rb
@@ -287,5 +287,48 @@ RSpec.describe TableStructure::Schema::Definition::Compiler do
         end
       end
     end
+
+    context 'when definitions contain nil' do
+      let(:definitions) do
+        [
+          nil,
+          { name: 'a' },
+          [nil, nil],
+          [nil, { name: 'b' }, nil]
+        ]
+      end
+
+      subject { described_class.new(name, definitions, options).compile }
+
+      context 'and `:nil_definitions_ignored` option is set `true`' do
+        let(:options) { { nil_definitions_ignored: true } }
+
+        it 'compiles definitions' do
+          expect(subject.size).to eq 2
+          expect(subject[0][:name]).to eq 'a'
+          expect(subject[1][:name]).to eq 'b'
+        end
+      end
+
+      context 'and `:nil_definitions_ignored` option is set `false`' do
+        let(:options) { { nil_definitions_ignored: false } }
+
+        it { expect { subject }.to raise_error TableStructure::Schema::Definition::Error }
+      end
+    end
+
+    context 'when definitions are empty' do
+      let(:definitions) do
+        [
+          []
+        ]
+      end
+
+      subject { described_class.new(name, definitions, options).compile }
+
+      it 'compiles definitions' do
+        expect(subject.size).to eq 0
+      end
+    end
   end
 end